+++ tests/testapp/test_listtextfield.py
@@ -258,7 +258,7 @@ class TestCheck(SimpleTestCase):
     def test_invalid_base_fields(self):
         class InvalidListTextModel2(TemporaryModel):
             field = ListTextField(
-                models.ForeignKey('testapp.Author'),
+                models.ForeignKey('testapp.Author', on_delete=models.CASCADE),
                 max_length=32
             )